ai-kb-generator/requirements.txt
flask>=3.0.3
flask-cors>=4.0.1
openai>=1.35.3
scikit-learn>=1.5.0
numpy>=1.26.4
nltk>=3.8.1
pandas>=2.2.2
python-dotenv>=1.0.1
tiktoken>=0.7.0
